refactor(backend): extract document preview text helpers
- move uploaded-document line cleanup, title selection, marker parsing, source-page parsing, and goal shaping into direct_document_preview_text.py
- keep legacy imports available through document preview payloads
- add focused text helper regression tests
- update cleanup audit and codebase navigation docs
