+++ docs/operations/WIII_RUNTIME_CLEANUP_AUDIT_2026-05-19.md
@@ -6,7 +6,7 @@ Owner: Project leadership
 
 Issue: #411
 
-Follow-up issues: #413, #415, #417, #419, #421, #423, #425, #427, #429, #431, #433, #435, #437, #439, #441, #477, #479, #481, #483, #485, #487 (owner: Architecture Maintainers)
+Follow-up issues: #413, #415, #417, #419, #421, #423, #425, #427, #429, #431, #433, #435, #437, #439, #441, #477, #479, #481, #483, #485, #487, #489, #491 (owner: Architecture Maintainers)
 
 ## Purpose
 
@@ -115,6 +115,10 @@ without broad deletes, secret exposure, or unreviewed product behavior changes.
   builders into `direct_prompt_tool_context.py`, while keeping graph-era
   compatibility exports and making query/state inputs explicit for best-effort
   skill prompt injection.
+- Follow-up #491 moved uploaded-document preview text shaping helpers into
+  `direct_document_preview_text.py`, keeping line cleanup, title selection,
+  marker extraction, learning-goal shaping, role-focused markdown, and
+  source-page parsing separate from host-action payload assembly.
 
 ## Preserved Intentionally
 
@@ -155,14 +159,15 @@ backups, data PDFs, or local skill folders.
   now live in focused helper modules.
 - `direct_tool_rounds_runtime.py` is now a smaller orchestration shell. Pointy,
   explicit web-search policy, deterministic document host-action execution,
-  uploaded-document preview/course payload shell, document source refs, domain
-  course plan builders, visual-turn policy, graph/runtime binding resolution,
-  per-turn provider policy, message builders, generic tool dispatch, final
-  synthesis helper construction, final synthesis execution, post-tool
-  convergence policy, follow-up LLM selection/invocation, response
-  finalization, post-tool search-template returns, and forced web-search
-  shortcuts have moved out. The next durable step is shrinking the compatibility
-  export surface once external imports can close.
+  uploaded-document preview/course payload shell, uploaded-document text
+  shaping, document source refs, domain course plan builders, visual-turn
+  policy, graph/runtime binding resolution, per-turn provider policy, message
+  builders, generic tool dispatch, final synthesis helper construction, final
+  synthesis execution, post-tool convergence policy, follow-up LLM
+  selection/invocation, response finalization, post-tool search-template
+  returns, and forced web-search shortcuts have moved out. The next durable
+  step is shrinking the compatibility export surface once external imports can
+  close.
 - `code_studio_template_scaffold.py` is still a large deterministic fallback,
   but contract, renderer dispatch, caption copy, and explicit-simulation
   quality policy are no longer embedded in the renderer body. Scene and
@@ -283,3 +288,9 @@ tool binding and tool-context prompt builders out of `direct_prompts.py`.
 Additional direct-node/guidance prompt tests passed with 69 tests. Targeted
 ruff checks, repository `ruff check app/ --select=E9,F63,F7`, and
 `git diff --check` also passed for the prompt tool-context extraction.
+In follow-up #491, focused document-preview text tests plus direct tool-round
+regressions passed with 86 tests after moving text cleanup/title/source-page
+helpers into `direct_document_preview_text.py`. The broader LMS/document
+preview contract regression set passed with 107 tests. Targeted ruff checks,
+repository `ruff check app/ --select=E9,F63,F7`, and `git diff --check` also
+passed for the text helper extraction.
